Add use of maintenance mode to admin procedure documentation
Not knowing the subtle and fine art of using `maintenance` mode when
temporarily taking content nodes down is a frequent occurrence amongst
users, which isn't too strange since you really have to know where
to look in order to learn about it. To make things a bit better, mention
the use of `vespa-set-node-state` explicitly under the ops procedures
for stopping and starting a Vespa node.